+++ src/smoothize.js
@@ -1,0 +1,5 @@
+function smoothize(from, to, ease) {
+  return from + (to - from) * ease;
+}
+
+module.exports = smoothize;